requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.3.0
debugpy==1.8.15
Flask==3.1.2
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.6.3
numpy==2.3.4
packaging==25.0
propcache==0.3.2
psycopg2-binary==2.9.11
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
